added laps table components and cleaned up styling
