+++ src/StopWatchButton.tsx
@@ -1,20 +1,11 @@
-import { Text, TouchableOpacity, View } from 'react-native';
+import { StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 
 export default function StopWatchButton(props: any) {
   const { title = '', onPress = () => {}, colour = '', isDisabled = false } = props;
   return (
-    <View style={{ margin: 10 }}>
+    <View style={{ margin: 30 }}>
       <TouchableOpacity
-        style={{
-            borderWidth:1,
-            borderColor: 'grey',
-            alignItems: 'center',
-            justifyContent:'center',
-            width: 100,
-            height: 100,
-            backgroundColor:colour,
-            borderRadius: 50,
-        }}
+        style={{ ...styles.button, backgroundColor: colour }}
         onPress={onPress}
         disabled={isDisabled}
       >
@@ -23,3 +14,15 @@ export default function StopWatchButton(props: any) {
     </View>
   );
 }
+
+const styles = StyleSheet.create({
+  button: {
+    borderWidth:1,
+    borderColor: 'grey',
+    alignItems: 'center',
+    justifyContent:'center',
+    width: 90,
+    height: 90,
+    borderRadius: 50,
+  },
+});
